Implement hiding of metadata in visualizations
**What**

Add option to `Visualizer` which tells it to hid metadata related
changes.

**Why**

Metadata has not been used so far in visualizations of deltas in
notifications. They could be however included if the original formatter
was invoked with the `showOperationalAttributes` (sometimes called
auxiliary attributes) set to true (e.g. from scripts).

In order to preserve the option to include/exclude metadata in delta
representation (together with operational attributes), the option to
hide metadata is added directly to the `Visualizer`, since it is
better place to handle it than it would be in a formatter itself.

**Implements**: MID-6112