Add API for wrapping specific parts of an overview
**What**

Add new API, which will allow (when integrated) to wrap some parts of
overview generated by the `Visualizer` and thus allow for partial
customization of the overview presentation. E.g. it can be used to add
bold font style to certain overview parts.

**Why**

Overview generated by `Visualizer` contains some specific parts, which
could potentially be handled in special way by the clients (e.g.
specific formatting may be applied).

Currently, such thing is not possible, because the overview is an
`LocalizableMessage` without any possibility of modifications. When
let's say UI wants to highlight certain part of the overview with bold
font style, then it needs to be set directly in the `LocalizableMessage`
thus directly in the localization resources as raw html tags.

There are also places, where overview could not be currently used (and
it isn't even generated for such cases), because the UI code for example
generates links from certain parts of the sentence, or adds various
icons. The new API which this commit adds, should allow to use overviews
(when integrated) also in such cases, where more custom presentation
of sentence parts is desired.

**Notes**

Technically speaking, the API allows not just wrapping of particular
parts of the localization to some arbitrary objects, but it also could
be used to totally transform them to something else. However that defies
the purpose of the interface and should never be used in such a way.

The reason, why localization parts should never be `transformed`, but
just wrapped is that the localization is basically defined by the
translator. Changing its individual parts, or changing the order of
parts in whole localization may make whole localization grammatically
incorrect or in worse case even change its meaning.