Fix computation of shadow items to return
If a client requested an item to be retrieved, and that item was not
returned by default, it sometimes did not get retrieved at all. The
logic in ShadowItemsToReturnProvider#shouldFetchExplicitly was
incorrect.

This commit fixes that, and resolves MID-10585.

(cherry picked from commit 36a5f0370a38b114b0bfee3b56bdb2a6d32b0184)