Fix solo correlators with composition item
If there was only a single correlator, it was used as a standalone one,
without (implicit) embedding composite correlator. Hence, its
"composition" configuration item (having e.g. the weight) was ignored.

This is now fixed. The reduction to solo correlator is done only if
there's no composition defined.

This resolves MID-10501.