Integrate localization parts wrapping API to Visualizations
**What**

Integrate the new API which allows to customize (wrap) parts of the
localization to the Visualizations abstraction as new kind of
"customizable overview".

**Why**

This new overviews implemented using the `WrapableLocalization` API
allows to do additional "formatting" of individual overview parts. That
means we can (eventually) get rid of the formatting html tags directly
in the localizations.

**Note**

Currently the new customizable overview is used only by the delta
formatters from notifications module. In order to use them also on the
UI, there are few necessary adjustments which needs to be done in the
`WrapableLocalizations`'s API (technically, it could be used also now,
but it would not be ideal).

The implementation of that API as it is right now, does translate the
underlying localization immediately after calling the `translate` method
on the `LocalizableObject` instance. Only after that translation, the
actual wrapping functions are applied.

However, in the Wicket UI, we would need to "postpone" the translation
to the invocation of the `getModelObject` method on the instance of the
`IModel` interface, which should be result of one of the wrappings.

In other words, the `translate` method on `LocalizableObject` should not
actually do the translation, but only provide some kind of a "supplier"
which would be then handed to the `IModel` instance, which in turn would
call that supplier when necessary.

The actual wrapping would then looks like this:
`translated localization supplier -> wrap to IModel -> wrap to ? extends
Component`. Only when the component would call the `getModelObject`
during the rendering phase, it would actually do the translation.

**Implements**: MID-6112