Stop loading ref. attributes if not needed
Provisioning module now honors EXCLUDE retrieval option for
ShadowType.F_ASSOCIATIONS: when present, it skips processing of
reference attributes and associations.

The primary use of this option is to speed up listing accounts from the
repository. But the post-processing can be turned off also for
operations that go to the resource.

Notes:

1. The use of F_ASSOCIATIONS is not quite precise, because we skip
also reference attributes. However, we currently do not have a simple
way how to specify "avoid processing all reference attributes".

2. This commit also fixes a problem when searching for shadows with
this option (no associations) destroyed cached associations' data in
the repository, because midPoint did not ask for them, and then naively
updated cached shadow. This is now fixed.

3. Turning off associations processing for fetching "getObject" is
currently only partial: reference attributes are fetched from the
resource (as it was up to now), only the post-processing is skipped.
We should fix that eventually.

Resolves MID-10444.

Unrelated change:

- Short and immediately obvious GetOperationOptions#noFetch method
was created instead of the original createNoFetchCollection() one.