Add "additional identification" formatter
**What**

Add implementation of formatter, which handles "additional
identification" properties, also known as "descriptive" properties.

**Why**

Descriptive properties in visualizations are properties, which
themselves were not changed, but they act as a hint about what
particular "container" was changed. The reason is, that sometimes the
mere name of the container is not enough.

**Notes**

Integration of this new formatter, is currently implemented as a
workaround and should be reworked later. The issues is, that currently
we can not rely on the `isDescriptive` method in the
`VisualizationItem`, because there is a bug (MID-10620), which causes
that `Visualizer` does not properly mark items as descriptive. In fact,
it does mark as descriptive also items, which are certainly not
descriptive. Another problem is that the UI sort of relies on that bug
(in fact, it was introduced as a "fix" for some UI bug).

**Implements**: MID-6112