Move bean deffinition of delta formatter to the xml
**What**

Move definition of `VisualizationBasedDeltaFormatter` bean to the xml.

**Why**

VisualizationBasedDeltaFormatter has many dependencies, some of which
does need string as a constructor parameter (e.g. indentation
character). For those classes it is not currently optimal to use
`@Component` annotation, because the configuration of mentioned string
parameters is not externalized. Thus I decided to put all the beans to
xml file.

**Implements**: MID-6112