Integrate new formatters to notifications
**What**

Integrate new formatters to the notifications

**Notes**

Old methods used to format deltas and objects which were in
`TextFormatter` has not had the `Task` and `OperationalResult`
parameters. But New implementation does need them in order to create
the `Visualization` instance. The issue is that former methods could be
potentially used also from scripts, so I couldn't just remove them. That
is the reason why they are still there. To workaround the need for task
and operational result, those methods call the `MidpointFunctions`
method to get the task.

The better approach, would be to separate the new methods (those which
requires task and operational result to new class, which would be called
from the original "task less" methods. The reason why it is not like
that right now is simply time.

**Implements**: MID-6112