Add mention of editing existing objType; convert filtering into section
Update the start of the tutorial to account also for cases when a user comes with an existing object type and wants to edit it; also moved the callout/mention of filtering to its own section because it's a rather important topic. A (related) shadow reclassification guide is TBD at this point.