update task GUIde + adjust limitations+panel config includes
This commit encompasses multiple loosely related changes: I updated the
Task GUIde with task naming examples and changed one wording to reflect
the GUIde is no longer about import tasks only (done in
3c411cf524b4b472007f21961f832076cee7ddf6 ). Furthermore, I removed
includes of the configuration panels and lifecycle summary in the
resource object wizard GUIdes because they are superfluous and
distracting. Instead, I linked the relevant articles. Lastly, I polished
the limitations-all.adoc and used that one in all the resource wizard
articles (except for correlation which is handles in another branch
currently) to make the limitations easier to manage, among other things.