Adapt schema tests
After introducing extension for GenericObjectType,
some schema tests started failing because of the change
of type name of the extension container. (This behavior
may or may not be 100% correct, because there could be
more extensions for one object type, but it is this way
for now.)

This commit adapts the tests.