Change smart integration service invocations
Up to now, ServiceClient was responsible for both HTTP communication
with the Python microservice and for conversion of requests and
responses between midPoint structures and the microservice API.

This is now changed: the conversion is done by ServiceAdapter,
and the ServiceClient takes care of HTTP communication only.

This allows e.g. better unit testing of the conversion code.