Add support for suggesting Groovy mappings
Now we call the Python microservice with a couple of example
data pairs to derive Groovy/as-is mappings.

Work in progress: inbounds only; no progress reporting.

Related change:

The serialization of values for Python microservice is switched
to xsd:string, with the specification of expected type. It is
because the midPoint-ish encoding of xsd:anyType is unnecessarily
complex for now (@type, @value).