Temporary fix to avoid Jekyll complaints about invalid link
The linked page exists currently only in master. Live web redirects the
link correctly but Jekyll doesn't work with htaccess files and thinks it
leads to 404 on <4.9, apparently (which it would, but web solves it byt
using master in the rendered link). So, until we make the structural
changes also in 4.9 and 4.8, I opted for this dirty fix with xrefv.