Add description to object class and attribute
This commit adds the "description" property for object classes
and "nativeDescription" property for attributes. (The name difference
is due to the fact that "description" for attributes already exists.)

It was added to midPoint schemas, ConnId interface, dummy connector,
and serialized (XSD) schemas.

To be polished later. For example, the attribute description is now
integrated with the configured variant (attributeDef.getDescription
returns either configured or native one), whereas for object classes
there is no such integration yet. The native description is visible
only via NativeObjectClassDefinition.

Cherry-picked and adapted from commits by Matus Macik:

- a3adf90859507a0a11d0fb1fb0fa40d14d520df6
- 8b115e94a1a99a3f8b869ad1f38dd17411544454
- 79736d9bc7d579b03330c0c68bf3e01eb3c19317