Mark AI-provided items as such
This commit introduces the concept of "Provided by AI" mark
on individual prism values. Currently, it is implemented by
having a provenance with

    originRef = origin-artificial-intelligence OID

where that object is a new initial object (ServiceType).

The code for manipulating that metadata is in AiUtil class.

Current support is for the data returned by
SmartIntegrationService (all 4 methods).

Related change: For simplicity, the returned value of
suggestFocusType method was changed from QName to
FocusTypeSuggestionType (to handle metadata even on
real values, without the need to use PPVs).