Add API & skeleton for better suggestFocusType
We no longer assume that the object type we are suggesting
focus type for already exists in the resource. The new
variant of suggestFocusType method takes the whole
object type definition (the one that should be in the
resource but is not there yet) as its input.

Work in progress - to be finished by Mišo.