Implement new Graal Polyglot based script evaluator
**What**

Implement new script evaluator for JavaScript, which is based on the Graal Polyglot APIs and GraalJS engine.

**Why**

We would like to use more secure alternatives to Groovy scripts for scripting. Graal Polyglot APIs, especially with the GraalJS
JavaScript engine seems to be a good fit for this because of theirs strong sandboxing capabilities.

**Notes**

This commit implements only the script evaluator. Other pieces and interfaces will be implemented in upcoming commits.

**Implements:** MID-10061