Provide skeleton for descriptive item paths
This commit lays the foundation of using so-called descriptive
item paths, e.g. c:email[*]/value. These paths indicate whether
a non-leaf item is single- or multivalued. We believe it's useful
for LLMs to reason about such paths.

This commit also re-structures the WP2/3 smart integration service
code by grouping it around individual operations, not layers of
abstraction. The primary motivation was that we need to keep the
state of such operations (to convert to/from descriptive item paths).

Related to MID-10840.