Add scripting language option to performance test
**What**

Add a configuration option to the performance test, which allows to
define what scripting language should be used.

**Why**

I used the performance test to measure performance of the midPoint using
GraaJS as an scripting engine for JavaScript. To do that I had to modify
the test xml files to contain correct `<language>` tag and in some cases
also modify the script itself to be a valid JavaScript. It would be a
shame to lost that work, so I decided to parameterize the test and allow
to specify what scripting language to use.