Adapt failing tests
TestSmartIntegrationServiceStory: this is just an adaptation
to the new "suggest object types" functionality, which
includes also focus type suggestion now.

TestIntegrationObjectWrapperFactory: adapted to the addition
of "complex attributes" (email, phone, address), although
this is to be reconsidered. They are containers, but should
they be represented in GUI as traditional containers?