Support "privileges" for scripting policy actions
The ScriptExecutionPolicyActionType.privileges item was
not used at all (by omission). This is now fixed.

There might be a slight performance degradation here:
originally, if there were multiple policy actions with
runAsRef pointing to specific user, they could be run
with just one login process (context switch). Unfortunately,
this is no longer the case, because of the implementation
aspects. See the improvement MID-10864.

Work in progress.

Related to MID-10820.