Fix opening shadow with object-side references
The problem manifested with the exception:

  "Only subject-side reference definition can have target
   object class definition"

occurring when a opening a shadow that contained a reference
attribute with the role of OBJECT. (Like "members" attribute
in a group.)

It stemmed from incomplete implementation of
ShadowReferenceAttributeDefinition#createTargetObjectsFilter.

This is now preliminarily fixed by - at least - not crashing
while executing the method.