Skip mapping suggestion if no target data
When suggesting mapping script for a particular target (focus) item,
we could go into situation that there is no data in that item.
This commit ensures that we don't attempt to derive the mapping script,
as It would yield "null" anyway.

Instead, we go for "asIs" mapping, assuming that the data will be
provided by the resource in question.

This commit also improves diagnostics logging for the operation.