Implement evaluation of expression generated by LLM
**What**

Implement evaluation of script expression, which was suggested by the
LLM.

**Why**

Scripts which are suggested by the LLM needs to be assessed for quality.
For such assessment we need to evaluate the script with certain input
and look to the result.

**Notes**

Current implementation hard-codes the expression profile with strict
permissions allowing only script evaluators. This should however be done
by some other means, not by hard-coding in the evaluation itself.