Fix inclusion of extensions in focus stats export
**What**

Fix the issue with unmatched extensions properties to include in focus
stats export.

**Note**

The main problem was, that the definition does not contain whole path,
but just the last segment (item name). When the include precondition was
applied, it did not matched any extension property, because the
precondition expected that the path will start with `extension/`. The
fix was to use `Item` instead of `ItemDefinition` in the precondition.
Item is aware of it's parents, so it contains whole path, not just item
name.

(cherry picked from commit df10f572fc2e3060b3641d8f8a3956c23887b1d1)