Change top level structure of shadow stats export
**What**

Change the "wrapper" for shadows stats export from the "generic" object
with extension to custom elements.

**Why**

The generic "wrapper" was used because of the rush. The custom elements
makes more sense from the "meaning" point of view. It is also better
aligned with the structure which is used for focus stats export.