Remove filter option from Shadow stats export command
**What**

Remove the `-f`, `--filter` options from the Export Shadow Statistics
command.

**Why**

The filter option is not relevant in this command, because there is
basically nothing to filter.

The aggregation function uses 5 shadow properties for grouping. In order
for statistics to work, we need all of those properties.

Values of these properties are not exported anyway, they are just used
for grouping.