Fix the language url in the permission profile
**What**

Fix the language url used in the permission profile and reduce the
permissions by using deny by default for evaluators and scripts.

**Why**

The default decisions were changed to ALLOW by default in previous
commit, because the evaluation failed in our tests. But the problem was
not in the default decision, but in the scripting language name used in
the profile.

The initial value of the language was not correct. In reality it has to
be an URL in specific form as you can see in the new constant.

Switching back to DENY by default on the evaluators level causes means
that no other than explicitly mentioned scripting evaluator will be
usable. That means no other expression types are allowed by this
profile.

Switching back to DENY by default on the scripting profile level causes
that only listed scripting languages (in our case Groovy) will be
allowed.