Support activity policy rules anywhere in tree
Now we can specify policy rules anywhere in the activity tree. They
will be evaluated in all child activities. For performance reasons,
the state (e.g., counters) is maintained locally, in each activity
involved. The initial values for each activity are determined at the
beginning, see PreexistingValues class.

Limitations:

1. Some tests in TestActivityPolicies fail.

2. There is still the problem of collecting rules from predefined
composite activities (like reconciliation) twice.

3. Policy state recording is done at the place where the policy rule
was triggered. But it should be at the place where it is defined, e.g.,
to treat notification suppression when the rule is triggered at
different places in the tree.

Related changes:

1. Unused type parameter in ActivityStateDefinition was removed.

2. Rule collecting is now done in LocalActivityRun, as it is not needed
for delegating and distributing activities. We can add it there later,
if necessary. The collecting was also removed from
ActivityPolicyProcessorHelper, as it's not needed there as well.

3. ActivityPolicyRuleIdentifier was added, to clearly mark this data
type.

4. Terminology change: instead of "enforcing" rules we "execute" them.
This is because rule enforcement is only one of possible actions in
object-based policy rules in midPoint. (Note: not all places were
changed yet.)

5. Minor code/doc improvements.