Fix support of multinode activity skip/restart
Support for activity aborting in multinode (worker) tasks was missing.
This is now fixed: "abortingWorkerRef" item was added, so worker tasks
can signal the aborting event to their coordinator.

DistributingActivityRun was adapted to support skip/restart feature.

Now TestActivityPolicies passes, except for test255 that is disabled.
We need to implement automatic worker abort when another worker is
aborting to make it pass.

Related to MID-10412.