Add resource based provider of correlation definition
**What**

Implement the provider of a correlation definition which reads the
definition from the resource.

**Why**

For "fast" simulation of correlations, we want to add an option for
including also existing correlation definition into the simulation. For
that purpose we need to read the correlation definition from the
resource. The work definition of the correlation simulation activity may
also contain correlation definition (in fact that is the main point of
implementing the activity).

That, however, means we have at least two sources of correlation
definitions for the activity. How those sources are accessed and
processed can differ significantly. For example when we read correlation
definition from the resource, we need to "merge" correlations, which are
possibly defined by different structures on different places of the
resource.

Because of that this change introduce the "provider" abstraction which
allows to encapsulate logic specific to particular source of
correlations.

**Notes**

As mentioned above, the correlation in resource can be defined in more
ways. Some of them are legacy, but we still support them. The provider
of the correlation implements a logic which collects all those
structures and merges them to single correlation definition. To do that
it relies on `CorrelatorsDefinitionUtils`, which provides necessary
functionality.