Fix state cleanup on activity restart
When activity is restarted, the existing state has to be cleaned up.
It is similar to repeating realization of an already completed activity.

The existing solution was simple but has the following issues:

1. when the state resided in subtasks, they were simply deleted,
2. the state persistence was not taken into account,
3. tree state overview was not updated.

Now the solution is more elaborate and integrated into
ActivityTreePurger, which was improved also by covering properties
that were added to the state since its inception (not related to
the activity restart functionality).

Also, experimental "restartCounters" property for "restartActivity"
policy action was removed. It was not necessary and the new code could
be written in a simpler way without it.

Related to MID-10412.