Implement union of correlation definition providers
**What**

Implement functionality to unite two providers of correlation
definition.

**Why**

In correlation simulation use case, the user can specify two sources of
correlation definition. One source is existing definition in a resource
and second source is inline definition in the work definition itself.
These sources are handled by correlation definition providers.

For the simulation, we need to merge definitions from all sources and
use the resulting definition.

To simplify that we can now "unite" the providers themself what creates
new instance of a provider which does the merging.

**Notes**

Currently the definition from provider reading definitions in resource
is used as a "target" for the merge operation. That has a consequence
mainly for the possibly present thresholds. If thresholds are in both
target and source (in this case inline definition in work def), only the
target threshold will be present in the result.

**Implements:** MID-10965