Fix approval delta classifications
Classification of deltas (related to approvals) into "applied",
"being applied", "waiting to be applied", "rejected" and similar
categories takes part of the information from the execution task.

If the task was deleted in the meanwhile, all applied changes were
erroneously displayed as "Waiting to be applied".

This commit fixes that by treating missing task situation appropriately:
it checks the case "closed" status to know the difference between
"waiting to be applied" and "applied" states.

This resolves MID-10828.