Ignore timed actions for cancelled cases
When a timed action was triggered for a cancelled approval case,
it could fail with "No outcomes?" illegal state exception. The reason
is that executing an action against already canceled case makes little
or no sense.

Hence, this commit makes sure that all timed actions are ignored for
cases that are already closed.

Related to MID-10934.