Implement correlation with specified correlation definition
**What**

Implement new method to the correlation service, which correlate given
shadow with given correlation definition.

**Why**

In the correlation simulation use case we need to simulate given
correlation definition on specified set of shadows. The correlation
definition to use does not have to be stored in the resource, but can
also be specified explicitly.

**Implements:** MID-10965