Integrate Correlation Activity to UI layer
**What**

Add code necessary for integration of Correlation Activity to the UI
layer.

**Why**

We want to trigger Correlation Activity task from the UI page which
shows correlations.

**Notes**

To add new activity task, I have modified the `ResourceTaskCreator` to
allow customization of the work definition. It now also allows to define
the query, which should be used to filter the resource objects.

I have also created the `ResourceTaskFlavors` class, which covers
activities (tasks) run on resource objects. It also includes the
"flavors" present in the `SynchronizationTaskFlavor` enum.

This new class also allows to find flavors by its archetype ID, what
simplifies the `ResourceTaskCreator`.

The intent is
to use the `ResourceTaskFlavors`, with the flavors defined as constants,
instead of the `SynchronizationTaskFlavor` enum from now on.