Fix validation of simple outbound mapping scripts
**What**

Fix the problem with undeclared variables when evaluating scripts from
outbound mappings.

**Why**

Outbound mappings works a differently than inbound mappings in a way
that the source for the transformation has the name of the property
itself. In case of inbound mapping it the name is always "input".

That of course caused problem when we tried to evaluate the outbound
mapping scripts, because they are not using the `input` variable, but
they use the variable which name is the same as the name of source
property.

**Notes**

This implementation is quite naive and it will work only in case of one
to one mappings. If the mapping contains more source properties, then it
will not work, because only one of them will be bound to script
variables map.